+++ src/Ext/BuildingType/Body.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <BuildingTypeClass.h>
+#include <SuperClass.h>
 #include <SuperWeaponTypeClass.h>
 
 #include <Helpers/Macro.h>
@@ -70,6 +71,10 @@ public:
 			, PlacementPreview_Palette {}
 			, PlacementPreview_Translucency {}
 		{ }
+
+		int GetSuperWeaponCount() const;
+		int GetSuperWeaponIndex(int index, HouseClass* pHouse) const;
+		int GetSuperWeaponIndex(int index) const;
 
 		virtual ~ExtData() = default;